requirements.txt
torch
torchvision
transformers
loguru
tqdm
scikit-learn
numpy
pandas
seaborn
matplotlib
PyYAML
nltk
